# docs/api-contract.yaml
openapi: 3.1.0
info:
  title: Ride-hailing API (Baghdad v1)
  version: 1.0.0
  description: |
    Binding API contract. CLAUDE.md §12.1: no endpoint may exist in the code that
    is not in this file. Adding an endpoint means changing this file first, in
    its own commit, with a reason.

    Conventions:
      * All money fields are named `*_iqd`, are integers, and are whole Iraqi
        Dinars (CLAUDE.md §6.1). A fractional value in any `*_iqd` field is a
        contract violation, not a rounding detail.
      * All timestamps are RFC 3339 UTC. Clients render in Asia/Baghdad.
      * All errors are RFC 9457 `application/problem+json` (CLAUDE.md §9).
      * Every request carries `X-Request-Id`; the server generates one if absent
        and always echoes it.
      * Bearer JWT auth unless the operation is explicitly marked `security: []`.
servers:
  - url: /v1
...
security:
  - bearerAuth: []
paths:
...
  /auth/otp/verify:
    post:
      tags: [auth]
      operationId: verifyOtp
      security: []
      summary: |
        Exchange a Firebase Phone Auth ID token for platform tokens.
        The OTP itself is sent and verified by Firebase on the device; the
        server never sees or generates an OTP code.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [firebaseIdToken, role]
              properties:
                firebaseIdToken: { type: string, minLength: 1, maxLength: 4096 }
                role:
                  type: string
                  enum: [RIDER, DRIVER]
                  description: |
                    ADMIN cannot be obtained here. Admin accounts are created by
                    another admin and sign in through the same endpoint with
                    role RIDER omitted — see /admin/session.
                displayName:
                  type: string
                  minLength: 1
                  maxLength: 120
                  description: Required only on first sign-up for role RIDER.
      responses:
        '200':
          description: Authenticated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthSession' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: |
            Role DRIVER requested but no driver record exists. In v1 admins
            create drivers manually (CLAUDE.md §2), so a driver cannot self
            sign-up.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/refresh:
    post:
      tags: [auth]
      operationId: refreshSession
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string, minLength: 1 }
      responses:
        '200':
          description: New token pair. The presented refresh token is rotated and revoked.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthSession' }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [auth]
      operationId: logout
      responses:
        '204': { description: Refresh tokens for this user revoked. }
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /devices:
    post:
      tags: [shared]
      operationId: registerDevice
      summary: Register this device for push notifications.
      description: >
        CONTRACT CHANGE, 2026-08-23. Added because push delivery had no way to
        reach a device: the backend enqueued push jobs and the worker logged
        and dropped them, and there was nowhere for a device token to be stored
        in the first place.

        Idempotent by token. Re-registering the same token refreshes its
        last_seen_at rather than creating a duplicate, because the apps
        re-register on every launch and FCM reissues tokens on reinstall.

        A token is bound to the authenticated caller. Registering a token that
        already belongs to another user REASSIGNS it - that is the correct
        behaviour when two people share a handset, and it is what stops ride
        offers following the previous owner.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, platform]
              properties:
                token:
                  type: string
                  minLength: 1
                  maxLength: 4096
                  description: The FCM registration token.
                platform:
                  type: string
                  enum: [ANDROID, IOS]
      responses:
        '204':
          description: Registered.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'
    delete:
      tags: [shared]
      operationId: unregisterDevice
      summary: Stop sending push notifications to this device.
      description: >
        Called on sign-out. Without it a driver who logs out keeps receiving
        ride offers on a phone they are no longer working from.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string, minLength: 1, maxLength: 4096 }
      responses:
        '204':
          description: Unregistered, or was not registered.
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me:
    get:
      tags: [shared]
      operationId: getMe
      responses:
        '200':
          description: The authenticated user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags: [shared]
      operationId: updateMe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName: { type: string, minLength: 1, maxLength: 120 }
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'

  # ===========================================================================
  # RIDER
  # ===========================================================================
...
  /me/capabilities:
    get:
      tags: [auth]
      operationId: getCapabilities
      summary: |
        What this account may do, decided by the server (CLAUDE.md §1.1).

        ALY is one app with a Rider mode and a Driver mode, so the client must
        not decide which it may enter. This returns the authoritative answer and
        is the SAME evaluation the server enforces with - if this and the guard
        could disagree, the app would offer a mode the server is about to
        refuse.

        Returns the FULL blocker list, not the first failure: a driver who is
        unapproved and also missing a licence needs to see both, or they fix one
        and learn the second only by failing again.
      responses:
        '200':
          description: The caller's capabilities.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Capabilities'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Missing or invalid token.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    ValidationError:
      description: Request failed schema validation.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    DriverDocumentType:
      type: string
      enum: [NATIONAL_ID, DRIVING_LICENCE, VEHICLE_REGISTRATION, VEHICLE_AUTHORIZATION]
...
    CapabilityBlocker:
      type: string
      description: |
        Why a mode is unavailable, machine-readable. The app localises these
        (CLAUDE.md §8); the server never returns a translated sentence.

        A client MUST tolerate an unknown value here rather than crashing - the
        server may add a blocker before the app is updated, and the safe
        degradation is a generic "contact support" row.
      enum:
        - ACCOUNT_DISABLED
        - NOT_A_DRIVER
        - APPROVAL_PENDING
        - APPROVAL_REJECTED
        - SUSPENDED
        - DOCUMENTS_INCOMPLETE
        - SUBSCRIPTION_REQUIRED
    DriverCapability:
      type: object
      required: [allowed, blockers, missingDocuments, expiredDocuments, rejectedDocuments]
      properties:
        allowed:
          type: boolean
          description: True only when blockers is empty. Never set independently.
        blockers:
          type: array
          items:
            $ref: '#/components/schemas/CapabilityBlocker'
        suspendedReason:
          type: string
          nullable: true
          description: |
            The only free text here, and the one deliberate exception to "no
            prose from the server": it is what an administrator actually typed,
            and a driver is entitled to read the reason they were given rather
            than a generic refusal.
        missingDocuments:
          type: array
          items:
            $ref: '#/components/schemas/DriverDocumentType'
        expiredDocuments:
          type: array
          items:
            $ref: '#/components/schemas/DriverDocumentType'
        rejectedDocuments:
          type: array
          items:
            $ref: '#/components/schemas/DriverDocumentType'
        subscriptionExpiresAt:
          type: string
          format: date-time
          nullable: true
          description: |
            Returned whether or not a subscription is currently required, so a
            driver can see what they bought before the policy is switched on.
    Capabilities:
      type: object
      required: [userId, canRide, driver]
      properties:
        userId:
          type: string
          format: uuid
        canRide:
          type: boolean
          description: |
            True for any account that is not disabled. There is no "rider
            approval" - refusing a banned account is the whole of the rider
            rule, and inventing more would be inventing policy.
        driver:
          $ref: '#/components/schemas/DriverCapability'
    Problem:
      type: object
      description: RFC 9457 problem details.
      required: [type, title, status]
      properties:
        type:     { type: string, format: uri }
        title:    { type: string }
        status:   { type: integer }
        detail:   { type: string }
        instance: { type: string }
        requestId: { type: string }
        errors:
          type: array
          description: Field-level validation failures.
          items:
            type: object
            required: [path, message]
            properties:
              path:    { type: string }
              message: { type: string }
      additionalProperties: true
...
    AuthSession:
      type: object
      required: [accessToken, refreshToken, expiresIn, user]
      properties:
        accessToken:  { type: string }
        refreshToken: { type: string }
        expiresIn:    { type: integer, description: Access token lifetime in seconds. }
        user:         { $ref: '#/components/schemas/Me' }
    Me:
      type: object
      required: [id, role, displayName, phone]
      properties:
        id:          { type: string, format: uuid }
        role:        { type: string, enum: [RIDER, DRIVER, ADMIN] }
        displayName: { type: string }
        phone:
          type: string
          description: |
            The caller's OWN phone, in E.164. Never present on any other user's
            representation — see PublicUser.
        rating:      { type: number, nullable: true, minimum: 1, maximum: 5 }
        driver:      { $ref: '#/components/schemas/DriverState' }
...
    DriverState:
      type: object
      required: [availability, isSuspended]
      properties:
        availability: { type: string, enum: [OFFLINE, ONLINE, ON_TRIP] }
        isSuspended:  { type: boolean }
        vehicle:
          type: object
          properties:
            plate: { type: string }
            model: { type: string }
            color: { type: string }
...